Module div's are now constructed based on device screen size for responsiveness
In mobile devices and portrait view the images do not appear and their positions in the screen are fit to be responsive. Same with the html canvas for map display.

Fixed bugs that occured when multiple events were fired from the user
i.e multiple clicks in the search bar would re produce a canvas draw function or the canvas was not properly cleared.